[GUI] alle Aktionen mit Fassadenaufrufen asynchron ausführen, um Anwendung nicht zu blockieren
